refactor(warp): resolve several quality code issues
- move warp node into the nodes folder
- create the proc module in nodes folder
- limit the deepcopy overwrite to the ObjectPayload only
- move remote_context in its own module
- improve elif and fix other minor code issue in remote service
- add a new dependency group in pyconfig.toml
