+++ juturna/nodes/__init__.py
@@ -1,6 +1,7 @@
 # noqa: D104
 from juturna.nodes import source
 from juturna.nodes import sink
+from juturna.nodes import proc
 
 
-__all__ = ['source', 'sink']
+__all__ = ['source', 'sink', 'proc']
