+++ pyproject.toml
@@ -40,6 +40,12 @@ fastapi-cli = { version = ">=0.0.7,<0.0.8", optional = true, extras = ["httpwrap
 prompt_toolkit = { version = "^3.0.52", optional = true, extras = ["pipebuilder"] }
 rich = { version = "^13.8.1", optional = true, extras = ["pipebuilder"] }
 
+# warp dependencies
+grpcio = { version = "1.76.0", optional = true, extras = ["warp"] }
+grpcio-tools = { version = "1.76.0", optional = true, extras = ["warp"] }
+protobuf = { version = "6.33.1", optional = true, extras = ["warp"] }
+typing_extensions = { version = "4.15.0", optional = true, extras = ["warp"] }
+
 # dev dependencies
 ruff = { version = "^0.14.8", optional = true, extras = ["dev"] }
 bandit = { version = "^1.9.2", optional = true, extras = ["dev"] }
@@ -67,6 +73,11 @@ authors = [ "dev@meetecho.com" ]
 [tool.poetry.extras]
 httpwrapper = ["fastapi", "fastapi-cli"]
 pipebuilder = ["prompt_toolkit", "rich"]
+warp = ["grpcio",
+        "grpcio-tools",
+        "protobuf",
+        "typing_extensions"
+        ]
 dev = ["pytest",
        "pre-commit",
        "bandit",
